0.0.6.230611c

+++ include/xq/net/rux_client.hpp
@@ -132,7 +132,7 @@ private:
         while (sockfd_ != INVALID_SOCKET) {
             n = recvfrom(sockfd_, (char*)frm->raw, sizeof(frm->raw), 0, (sockaddr*)&frm->name, &frm->namelen);
             if (n < 0) {
-                ev_.on_error(ErrType::IO_RCV, (void*)errcode);
+                ev_.on_error(ErrType::IO_RCV, (void*)((int64_t)errcode));
                 continue;
             }
 
@@ -182,7 +182,7 @@ private:
             for (i = 0; i < n; i++) {
                 frm = frms[i];
                 if (::sendto(sockfd_, (char*)frm->raw, frm->len, 0, (sockaddr*)&frm->name, frm->namelen) < 0) {
-                    ev_.on_error(ErrType::IO_SND, (void*)errcode);
+                    ev_.on_error(ErrType::IO_SND, (void*)((int64_t)errcode));
                 }
                 delete frm;
             }
@@ -237,7 +237,7 @@ private:
             if (n < 0) {
                 err = errcode;
                 if (err != EAGAIN && err != EINTR) {
-                    ev_.on_error(ErrType::IO_RCV, (void*)errcode);
+                    ev_.on_error(ErrType::IO_RCV, (void*)((int64_t)errcode));
                     break;
                 }
                 continue;
@@ -318,7 +318,7 @@ private:
                 }
 
                 if (res < 0) {
-                    ev_.on_error(ErrType::IO_SND, (void*)errcode);
+                    ev_.on_error(ErrType::IO_SND, (void*)((int64_t)errcode));
                 }
             }
         }
